DB connection refused
Still trying to get this working. From a bash terminal inside the DB container the login information of root and root work, but Installing nmap on the project's container the 3306 port exists under mysql but is closed. Not sure why. The h2 driver is using the correct credentials but says that the connection was refused and obviously it is closed. Docker-compose should leave it open.